Deleted immutability on object
The nice things immutability brings as a pattern are not enough in
front of the bad things it brings
- This request is, in fact, a non-immutable thing per concept. Different
layers can complete the object by adding key-values, so the class should
be open for contributions.
- When working with listeners or subscribers, these should be able to
wotk with this obect, adding tracebility for example. Working with this
immutability makes incredible difficult to do something like this
without adding a method like `updateContext` on Request
- Coupling between layers in terms of object is even normal. The contect
is shared between all of them, and the context should be updated as long
as is updated itself. Because values are indeed immutable, at least in
terms of key, we should'nt have any problem here